Rasterisation of multiple cubes
This rendering design isn't very efficient because it relies on atomic depth buffer writes, and the load for different workgroups varies significantly, as cubes that are closer to the camera take up much more surface area. There are no guardrails in place to make sure that the same thread that successfully writes to the depth buffer also writes to the colour texture, thus resulting in noticeable graphical glitches.